calcTaskDetails moved to util
